requirements.txt
Babel==2.3.4
beautifulsoup4==4.4.1
boto==2.40.0
csvkit==0.9.1
dbf==0.94.3
Delorean==0.6.0
Django==1.9
django-admin-bootstrapped==2.5.7
django-bakery==0.8.4
django-debug-toolbar==1.4
django-mass-edit==2.6
django-tastypie==0.13.3
ecdsa==0.13
Fabric==1.11.1
humanize==0.5.1
itsdangerous==0.24
jdcal==1.2
latimes-calculate==0.3.1
lxml==3.6.0
MarkupSafe==0.23
MySQL-python==1.2.5
newrelic==2.66.0.49
openpyxl==2.2.0b1
paramiko==1.16.0
pycrypto==2.6.1
python-dateutil==2.5.3
python-googlegeocoder==0.3.0
python-mimeparse==1.5.2
python-slugify==1.2.0
pytz==2016.4
PyYAML==3.11
redis==2.10.5
requests==2.9.1
six==1.10.0
slacker==0.9.9
slacker-log-handler==1.2.0
SQLAlchemy==1.0.12
sqlparse==0.1.19
twitter==1.15.0
tzlocal==1.2.2
Unidecode==0.4.19
Werkzeug==0.11.10
xlrd==0.9.4
